requirements.txt
ctransformers 0.2.27
sentence-transformers==2.2.2
pinecone-client
langchain==0.0.225
flask
pypdf
ipykernel
pathlib
python-dotenv
-e .